pyproject.toml
[project]
name = "doan"
dynamic = ["version"]
authors = [{ name = "Lâm Quang Trí", email = "quangtri.lam.9@gmail.com" }]
maintainers = [{ name = "Lâm Quang Trí", email = "quangtri.lam.9@gmail.com" }]
description = "Hệ thống hỏi đáp sản phẩm điện tử"
license = "MIT"
license-files = ["LICENSE"]
readme = "README.MD"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai",
    "langchain-community",
    "qdrant-client",
    "sentence-transformers",
    "python-dotenv",
    "torch",
    "pandas",
    "numpy",
    "pydantic",
    "pydantic-settings",
    "langchain-huggingface",
    "langchain-docling",
    "typer",
    "rich",
    "langchain-qdrant",
    "structlog",
    "streamlit",
    "duckduckgo-search",
    # API dependencies
    "litestar[brotli,cryptography,jwt,standard,structlog,redis]",
    "litestar-granian",
    "bcrypt",
    "md2docx-python",
    "aiohttp",
    "ulid",
    "pytest>=8.4.0",
    "datasets>=3.6.0",
]

[tool.ruff.lint]
extend-select = ["CPY"]
preview = true
